Combined MigrationService and TaxableMigrationService.
